Ajustes na recuperação de partidas

+++ personalstats/src/main/java/com/junior/personalstats/repository/MatchRepository.java
@@ -1,9 +1,11 @@
 package com.junior.personalstats.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.junior.personalstats.model.Match;
 
-public interface MatchRepository extends MongoRepository<Match, Long>{
+public interface MatchRepository extends MongoRepository<Match, BigInteger>{
 
 }
